requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
google-generativeai
cryptography
Pillow
python-multipart
httpx   
mysql-connector-python
#pip install aiofiles
